# Convert the test suite to use transactional tests