--- a/core/cmd/local_client.go
+++ b/core/cmd/local_client.go
@@ -1,24 +1,31 @@
 package cmd
 
 import (
+	"database/sql"
 	"fmt"
 	"io"
 	"io/ioutil"
 	"math/big"
+	"net/url"
 	"os"
+	"path"
 	"path/filepath"
+	"runtime"
 	"strings"
 
 	"github.com/pkg/errors"
 
+	"github.com/smartcontractkit/chainlink/core/gracefulpanic"
 	"github.com/smartcontractkit/chainlink/core/logger"
 	"github.com/smartcontractkit/chainlink/core/services/chainlink"
 	strpkg "github.com/smartcontractkit/chainlink/core/store"
+	"github.com/smartcontractkit/chainlink/core/store/migrations"
 	"github.com/smartcontractkit/chainlink/core/store/models"
 	"github.com/smartcontractkit/chainlink/core/store/orm"
 	"github.com/smartcontractkit/chainlink/core/store/presenters"
 	"github.com/smartcontractkit/chainlink/core/utils"
 
+	"github.com/jinzhu/gorm"
 	clipkg "github.com/urfave/cli"
 	"go.uber.org/zap/zapcore"
 )
@@ -287,6 +294,100 @@ func (cli *Client) RebroadcastTransactions(c *clipkg.Context) error {
 		}
 	}
 	return nil
+}
+
+// ResetDatabase drops, creates and migrates the database specified by DATABASE_URL
+// This is useful to setup the database for testing
+func (cli *Client) ResetDatabase(c *clipkg.Context) error {
+	logger.SetLogger(cli.Config.CreateProductionLogger())
+	config := orm.NewConfig()
+	if config.DatabaseURL() == "" {
+		return cli.errorOut(errors.New("You must set DATABASE_URL env variable. HINT: If you are running this to set up your local test database, try DATABASE_URL=postgresql://postgres@localhost:5432/chainlink_test?sslmode=disable"))
+	}
+	logger.Infof("Resetting database: %#v", config.DatabaseURL())
+	if err := dropAndCreateDB(config); err != nil {
+		return cli.errorOut(err)
+	}
+	if err := migrateTestDB(config); err != nil {
+		return cli.errorOut(err)
+	}
+	return nil
+}
+
+// PrepareTestDatabase calls ResetDatabase then loads fixtures required for tests
+func (cli *Client) PrepareTestDatabase(c *clipkg.Context) error {
+	if err := cli.ResetDatabase(c); err != nil {
+		return err
+	}
+	config := orm.NewConfig()
+	if err := insertFixtures(config); err != nil {
+		return cli.errorOut(err)
+	}
+	return nil
+}
+
+func dropAndCreateDB(config *orm.Config) error {
+	parsed, err := url.Parse(config.DatabaseURL())
+	if err != nil {
+		return err
+	}
+
+	dbname := parsed.Path[1:]
+	// Cannot drop the database if we are connected to it, so we must connect
+	// to a different one. template1 should be present on all postgres installations
+	parsed.Path = "/template1"
+	db, err := sql.Open(string(orm.DialectPostgres), parsed.String())
+	if err != nil {
+		return fmt.Errorf("unable to open postgres database for creating test db: %+v", err)
+	}
+	defer db.Close()
+
+	_, err = db.Exec(fmt.Sprintf("DROP DATABASE IF EXISTS %s", dbname))
+	if err != nil {
+		return fmt.Errorf("unable to drop postgres database: %v", err)
+	}
+	// `CREATE DATABASE $1` does not seem to work w CREATE DATABASE
+	_, err = db.Exec(fmt.Sprintf("CREATE DATABASE %s", dbname))
+	if err != nil {
+		return fmt.Errorf("unable to create postgres database: %v", err)
+	}
+	return nil
+}
+
+func migrateTestDB(config *orm.Config) error {
+	orm, err := orm.NewORM(config.DatabaseURL(), config.DatabaseTimeout(), gracefulpanic.NewSignal(), config.GetDatabaseDialectConfiguredOrDefault(), config.GetAdvisoryLockIDConfiguredOrDefault())
+	if err != nil {
+		return fmt.Errorf("failed to initialize orm: %v", err)
+	}
+	orm.SetLogging(config.LogSQLStatements() || config.LogSQLMigrations())
+	err = orm.RawDB(func(db *gorm.DB) error {
+		return migrations.GORMMigrate(db)
+	})
+	if err != nil {
+		return fmt.Errorf("migrateTestDB failed: %v", err)
+	}
+	orm.SetLogging(config.LogSQLStatements())
+	return orm.Close()
+}
+
+func insertFixtures(config *orm.Config) error {
+	db, err := sql.Open(string(orm.DialectPostgres), config.DatabaseURL())
+	if err != nil {
+		return fmt.Errorf("unable to open postgres database for creating test db: %+v", err)
+	}
+	defer db.Close()
+
+	_, filename, _, ok := runtime.Caller(1)
+	if !ok {
+		return errors.New("could not get runtime.Caller(1)")
+	}
+	filepath := path.Join(path.Dir(filename), "../internal/fixtures/fixtures.sql")
+	fixturesSQL, err := ioutil.ReadFile(filepath)
+	if err != nil {
+		return err
+	}
+	_, err = db.Exec(string(fixturesSQL))
+	return err
 }
 
 // DeleteUser is run locally to remove the User row from the node's database.