--- a/core/internal/fixtures/apicredentials
+++ b/core/internal/fixtures/apicredentials
@@ -1,2 +1,2 @@
-email@test.net
+apiuser@chainlink.test
 password