Each instance of a store in tests now runs in its own database transaction.

This has several benefits:

- Significantly speeds up the test suite
- Makes the test suite more reliable by eliminating "cannot drop database that is in use" errors
- Doesn't leave rubbish databases hanging around after test suite has finished.

Time taken to run time `go test -p 8 ./... -parallel 8` on MacBook Pro 16" with 8 core CPU:

Before: 690.65s user 66.47s system 532% cpu 2:22.26 total
After: 336.88s user 24.13s system 560% cpu 1:04.38 total

In addition:

- Fixed many timing issues and subtly broken tests that I found during the course of implementing this PR.
- Introduced a mockable KeyStore
- Update the keys.address column to be citext to prevent duplicate keys.
- Introduce a new config env variable, `MIGRATE_DATABASE`, that enables or disables automatic database migrations on application start.

NOTE: Developers are now required to run an additional step (`go run . local db preparetest`) to prepare their database before tests. This does not need to be re-run on every test run, but will need to be rerun if the schema changes.

Our postgres tests should now be guaranteed to be 100% non-flaky 🤞