@samsondav 2020-04-30
This is now mostly there, needs a full self-review and cleanup of some dead ends I went down before.